Split mod_core and mod_loader into new files, include via include_str!
