+++ src/mod_loader.lua
@@ -1,0 +1,45 @@
+function G.UIDEF.mods()
+    btn_nodes = {}
+    for i, mod in ipairs(mods) do
+        col = G.C.RED
+        if mod.enabled then
+            col = G.C.GREEN
+        end
+        table.insert(btn_nodes, UIBox_button({
+            minw = 6,
+            button = "usage",
+            minh = 0.8,
+            colour = col,
+            label = {
+                mod.name
+            }
+        }))
+    end
+    return (create_UIBox_generic_options({
+        snap_back = true,
+        back_func = "options",
+        contents = {
+            {
+                n = G.UIT.C,
+                config = {
+                    r = 0.1,
+                    align = "cm",
+                    padding = 0.1,
+                    colour = G.C.CLEAR
+                },
+                nodes = btn_nodes
+            }
+        }
+    }))
+
+end
+
+function G.FUNCS.show_mods(e)
+    G.SETTINGS.paused = true
+
+    G.FUNCS.overlay_menu({
+        definition = G.UIDEF.mods()
+    })
+end
+
+G.VERSION = G.VERSION .. "\nBalamod {balamod_version}"
